stop injecting the container when possible
Injecting the container in your services is not a good practice. After
this, there are only three classes left that rely on the container being
injected: The `RestRouteLoader` class needs it to resolve controllers
registered as services, the `ParamFetcher` class injects the container
into container aware `ParamInterface` implementations, and the parent
class of the `ViewResponseListener` uses the container.